refactor: use uint64 consistently

+++ internal/server/log.go
@@ -9,7 +9,7 @@ var ErrOffsetNotFound = fmt.Errorf("offset not found")
 
 type Record struct {
 	Value  []byte `json:"value"`
-	Offset uint   `json:"offset"`
+	Offset uint64 `json:"offset"`
 }
 
 type Log struct {
@@ -21,18 +21,18 @@ func NewLog() *Log {
 	return &Log{}
 }
 
-func (l *Log) Append(record Record) (uint, error) {
+func (l *Log) Append(record Record) (uint64, error) {
 	l.mtx.Lock()
 	defer l.mtx.Unlock()
-	record.Offset = uint(len(l.records))
+	record.Offset = uint64(len(l.records))
 	l.records = append(l.records, record)
 	return record.Offset, nil
 }
 
-func (l *Log) Read(offset uint) (Record, error) {
+func (l *Log) Read(offset uint64) (Record, error) {
 	l.mtx.Lock()
 	defer l.mtx.Unlock()
-	if offset >= uint(len(l.records)) {
+	if offset >= uint64(len(l.records)) {
 		return Record{}, ErrOffsetNotFound
 	}
 	return l.records[offset], nil
